pyproject.toml
[tool.poetry]
name = "duz-game"
version = "0.1.0"
description = "Introducing Duz, a captivating Telegram app/game inspired by the traditional Persian game \"Duz.\" This project leverages the power of Python and PostgreSQL to deliver a more engaging and enjoyable experience than the classic game of Tic-Tac-Toe (XO)."
authors = ["amirhossein <amirhosseinmobini9@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.3"
python-dotenv = "^1.0.1"
watchdog = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
